requirements.txt
fastapi==0.101.0
uvicorn==0.23.0
PyJWT==2.8.0
aioredis==2.0.1
httpx==0.24.1
pytest==7.4.0
pytest-asyncio==0.22.0
pydantic==1.10.11
python-json-logger==2.0.4
